add toTransactionCompletedEvent mapper in Transaction entity file
