--- wallet/src/main/kotlin/net/thechance/wallet/eventListener/mapper/TransactionCompletedEventMapper.kt
@@ -1,2 +1,0 @@
-package net.thechance.wallet.eventListener.mapper
-
